Unpivot function
Hello! Here is the script for unpivoting the data frames. I think this will be useful prior to join tables and making relationships to create a data warehouse.
Since is my first time uploading a script to the project, I didn't know where to paste it, so I decided to upload it in a .txt file.